feat: rename route from billing to price
