+++ src/constant/config/route.ts
@@ -1,3 +1,3 @@
 export const AUTH_ROUTES = ['/login', '/signup'] as const;
 
-export const PROTECTED_ROUTES = ['/chat', '/billing'] as const;
+export const PROTECTED_ROUTES = ['/chat', '/pricing'] as const;
